perf(搜索): 使用HashSet优化搜索性能
将本地和服务器搜索结果中的ID存储从ArrayList改为HashSet，将查找复杂度从O(n)降为O(1)

+++ TMessagesProj/src/main/java/org/telegram/messenger/MediaDataController.java
@@ -3779,11 +3779,11 @@ public class MediaDataController extends BaseController {
         FileLog.d("wd 服务器搜索结果添加数量: " + serverAddedCount);
         
         //wd 搜索优化：使用HashSet替代ArrayList.contains()，将O(n)查找降为O(1)
-        HashSet<Long> localIds = new HashSet<>();
+        HashSet<Integer> localIds = new HashSet<>();
         for (int i = 0; i < searchLocalResultMessages.size(); i++) {
             localIds.add(searchLocalResultMessages.get(i).messageOwner.id);
         }
-        HashSet<Long> loadedIds = new HashSet<>();
+        HashSet<Integer> loadedIds = new HashSet<>();
         if (loadedMessages != null) {
             for (int i = 0; i < loadedMessages.size(); i++) {
                 loadedIds.add(loadedMessages.get(i).messageOwner.id);
